+++ yggdrasil_decision_forests/port/python/ydf/model/decision_forest_model.py
@@ -14,6 +14,8 @@
 
 """Definitions for generic decision forest models."""
 
+from typing import Optional
+
 import numpy as np
 
 from ydf.cc import ydf
@@ -50,3 +52,60 @@ class DecisionForestModel(generic_model.GenericModel):
         data, data_spec=self._model.data_spec()
     )
     return self._model.PredictLeaves(ds._dataset)  # pylint: disable=protected-access
+
+  def distance(
+      self,
+      data1: dataset.InputDataset,
+      data2: Optional[dataset.InputDataset],
+  ) -> np.ndarray:
+    """Computes the pairwise distance between examples in "data1" and "data2".
+
+    If "data2" is not provided, computes the pairwise distance between examples
+    in "data1".
+
+    Usage example:
+
+    ```python
+    import pandas as pd
+    import ydf
+
+    # Train model
+    train_ds = pd.read_csv("train.csv")
+    model = ydf.RandomForestLearner(label="label").Train(train_ds)
+
+    test_ds = pd.read_csv("test.csv")
+    distances = model.distance(test_ds, train_ds)
+    # "distances[i,j]" is the distance between the i-th test example and the
+    # j-th train example.
+    ```
+
+    Different models are free to implement different distances with different
+    definitions. For this reasons, unless indicated by the model, distances
+    from different models cannot be compared.
+
+    The distance is not guaranteed to satisfy the triangular inequality
+    property of metric distances.
+
+    Not all models can compute distances. In this case, this function will raise
+    an Exception.
+
+    Args:
+      data1: Dataset. Can be a dictionary of list or numpy array of values,
+        Pandas DataFrame, or a VerticalDataset.
+      data2: Dataset. Can be a dictionary of list or numpy array of values,
+        Pandas DataFrame, or a VerticalDataset.
+
+    Returns:
+      Pairwise distance
+    """
+
+    ds1 = dataset.create_vertical_dataset(
+        data1, data_spec=self._model.data_spec()
+    )
+    if data1 is None:
+      ds2 = ds1
+    else:
+      ds2 = dataset.create_vertical_dataset(
+          data2, data_spec=self._model.data_spec()
+      )
+    return self._model.Distance(ds1._dataset, ds2._dataset)  # pylint: disable=protected-access